refactor : add packages jwt, product 패키지 추가
