--- src/main/java/gift/ProductViewController.java
@@ -1,2 +1,0 @@
-package gift;public class ProductViewController {
-}
