+++ src/main/java/gift/product/Product.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.product;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
