+++ src/main/java/gift/product/ProductViewController.java
@@ -1,0 +1,2 @@
+package gift.product;public class ProductViewController {
+}
