+++ src/main/java/gift/user/UserController.java
@@ -1,15 +1,10 @@
 package gift.user;
 
 
-import gift.ResponseDTO;
-import io.jsonwebtoken.JwtException;
+import gift.jwt.JWTService;
 import jakarta.validation.Valid;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.HashMap;
 
 @RestController
 @RequestMapping("/members")
